make smoketest setup check for empty strings
Apparently terraform can give us output saying that it succeded,
but give us an empty string for an IP address, which doesn't actually
let us use the IP address. Check for that case too in our overly
fancy setup.

This is a precursor to figuring out what's going wrong in a way
that lets us fix it more properly.

Also, request values, but don't instant-exit if they aren't present,
so we can actually do the retries.